Only check for zero length files in fdupes mode
The file scan will skip any files whose size is less than our blocksize.
This makes sense if we're computing hashes but fdupes mode just sends the
entire file to the kernel for dedupe. So if we're in fdupes mode, only check
that the file has a size at all - zero length files should still be skipped.

Signed-off-by: Mark Fasheh <mfasheh@suse.de>